refactor: split public key handlers
refactor: move legacy version handlers
refactor: move legacy hwm handler
