+++ rust/app/src/handlers/legacy/public_key.rs
@@ -13,3 +13,68 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 ********************************************************************************/
+use std::convert::TryFrom;
+
+use crate::{
+    constants::ApduError as Error,
+    crypto,
+    dispatcher::ApduHandler,
+    handlers::public_key::{Addr, GetAddress},
+    sys::{self, Show},
+};
+
+pub struct LegacyGetPublic;
+pub struct LegacyPromptAddress;
+
+impl ApduHandler for LegacyGetPublic {
+    fn handle<'apdu>(
+        _: &mut u32,
+        tx: &mut u32,
+        buffer: crate::utils::ApduBufferRead<'apdu>,
+    ) -> Result<(), Error> {
+        *tx = 0;
+
+        //TODO: require_hid ?
+        // see: https://github.com/Zondax/ledger-tezos/issues/35
+
+        let curve = crypto::Curve::try_from(buffer.p2()).map_err(|_| Error::InvalidP1P2)?;
+
+        let cdata = buffer.payload().map_err(|_| Error::DataInvalid)?;
+        let bip32_path =
+            sys::crypto::bip32::BIP32Path::<6>::read(cdata).map_err(|_| Error::DataInvalid)?;
+
+        let key = GetAddress::new_key(curve, &bip32_path).map_err(|_| Error::ExecutionError)?;
+
+        let key = key.as_ref();
+        let len = key.len();
+        buffer.write()[..len].copy_from_slice(key);
+
+        *tx = len as u32;
+
+        Ok(())
+    }
+}
+
+impl ApduHandler for LegacyPromptAddress {
+    fn handle<'apdu>(
+        flags: &mut u32,
+        tx: &mut u32,
+        buffer: crate::utils::ApduBufferRead<'apdu>,
+    ) -> Result<(), Error> {
+        *tx = 0;
+
+        let curve = crypto::Curve::try_from(buffer.p2()).map_err(|_| Error::InvalidP1P2)?;
+
+        let cdata = buffer.payload().map_err(|_| Error::DataInvalid)?;
+        let bip32_path =
+            sys::crypto::bip32::BIP32Path::<6>::read(cdata).map_err(|_| Error::DataInvalid)?;
+
+        let key = GetAddress::new_key(curve, &bip32_path).map_err(|_| Error::ExecutionError)?;
+
+        let ui = Addr::new(&key)
+            .map_err(|_| Error::DataInvalid)?
+            .into_ui(key, false);
+
+        unsafe { ui.show(flags) }.map_err(|_| Error::ExecutionError)
+    }
+}
