+++ rust/app/src/handlers/mod.rs
@@ -13,7 +13,6 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 ********************************************************************************/
-pub mod legacy_version;
 pub mod public_key;
 pub mod signing;
 pub mod version;
@@ -24,9 +23,6 @@ pub mod legacy;
 
 #[cfg(feature = "dev")]
 pub mod dev;
-
-#[cfg(feature = "baking")]
-pub mod hwm;
 
 #[cfg(feature = "baking")]
 pub mod baking;
